requirements.txt
faster-whisper>=1.0.0
customtkinter>=5.2.0
tkinterdnd2>=0.3.0
imageio-ffmpeg>=0.4.9
huggingface_hub>=0.20.0
pillow>=10.0.0
pyinstaller>=6.0.0
# CUDA runtime DLLs — required for GPU acceleration on Windows.
# ctranslate2 does not bundle these; the nvidia packages provide them.
nvidia-cublas-cu12>=12.3
nvidia-cuda-runtime-cu12>=12.3
